# src/doc/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample ecommerce server. For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Ecommerce",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "waltongerman@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/cart": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Create a new cart",
        "description": "Create a new cart into the ecommerce database",
        "operationId": "createCart",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "The product object that needs to be added to the ecommerce",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cart"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation succesfull"
          }
        }
      },
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Get all carts",
        "description": "",
        "operationId": "getAllCarts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "schema": {
              "items": {
                "$ref": "#/definitions/Cart"
              }
            }
          }
        }
      }
    },
    "/cart/{cartId}": {
      "get": {
        "tags": [
          "cart"
        ],
        "summary": "Find a cart by id",
        "description": "Returns a single cart",
        "operationId": "getCartById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Id of the cart to find",
            "required": true,
            "type": "string",
            "items": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Cart not found",
            "schema": {
              "items": {
                "$ref": "#/definitions/Cart"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Delete a cart",
        "description": "",
        "operationId": "deleteCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Id of the cart that need to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operation succesfull"
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
...
    "/cart/{cartId}/{productId}": {
      "post": {
        "tags": [
          "cart"
        ],
        "summary": "Add a product in the cart with his id",
        "description": "Add a product in the cart",
        "operationId": "addToCart",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Id of the cart that you want to add a product",
            "required": true,
            "type": "string"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "Id of the product that you want to add to the cart",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Operation successfull"
          },
          "404": {
            "description": "Cart or product not found"
          }
        }
      }
    },
    "/cart/{cartId}/products/{prodCode}": {
      "delete": {
        "tags": [
          "cart"
        ],
        "summary": "Delete a product in the cart",
        "description": "",
        "operationId": "deleteCartProduct",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "cartId",
            "in": "path",
            "description": "Id of the cart",
            "required": true,
            "type": "string"
          },
          {
            "name": "prodCode",
            "in": "path",
            "description": "Id of the product you want to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Operation succesfull"
          },
          "404": {
            "description": "Product or cart not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Cart": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "xml": {
            "name": "product",
            "wrapped": true
          },
          "items": {
            "type": "string"
          },
          "default": []
        },
        "timestamp": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Cart"
      }
    },
...
  }
}